multa-api/requirements.txt
boto3==1.13.1
django==3.0.5
djangorestframework==3.11.0
python-decouple==3.3